✨ feat: 프로필 거래 상태 버튼 수정 / 프로필 수정 경로 변경 / 불필요한 코드 제거
