--- src/app/profile/[id]/page.tsx
@@ -1,3 +1,0 @@
-export default function Profile() {
-  return;
-}
